+++ crates/universal-sierra-compiler-api/src/lib.rs
@@ -9,7 +9,7 @@ use std::io::Write;
 use std::str::from_utf8;
 use tempfile::Builder;
 
-use crate::spinner::spawn_spinner_message;
+use crate::spinner::spawn_usc_spinner;
 pub use command::*;
 use shared::command::CommandExt;
 
@@ -72,25 +72,25 @@ pub fn compile_sierra_at_path<T: UniversalSierraCompilerOutput>(
     sierra_file_path: &Utf8Path,
     sierra_type: &SierraType,
 ) -> Result<T> {
-    let spinner = spawn_spinner_message(sierra_file_path)?;
+    let usc_output = {
+        let _spinner = spawn_usc_spinner(sierra_file_path)?;
 
-    let mut usc_command = UniversalSierraCompilerCommand::new();
-    let usc_output = usc_command
-        .inherit_stderr()
-        .args(vec![
-            &("compile-".to_string() + &sierra_type.to_string()),
-            "--sierra-path",
-            sierra_file_path.as_str(),
-        ])
-        .command()
-        .output_checked()
-        .context(
-            "Error while compiling Sierra. \
+        let mut usc_command = UniversalSierraCompilerCommand::new();
+        usc_command
+            .inherit_stderr()
+            .args(vec![
+                &("compile-".to_string() + &sierra_type.to_string()),
+                "--sierra-path",
+                sierra_file_path.as_str(),
+            ])
+            .command()
+            .output_checked()
+            .context(
+                "Error while compiling Sierra. \
             Make sure you have the latest universal-sierra-compiler binary installed. \
             Contact us if it doesn't help",
-        )?;
-
-    spinner.finish_and_clear();
+            )?
+    };
 
     Ok(T::convert(from_utf8(&usc_output.stdout)?.to_string()))
 }